refactored app to move functionality to useContext
