+++ src/components/shared/Modal.jsx
@@ -1,9 +1,13 @@
 import React from "react";
 import Button from "./Button";
+import ProductivityContext from "../../context/ProductivityContext";
+import { useContext } from "react";
 
-const Modal = ({ setModalOpen, deleteEntry, entryId }) => {
+const Modal = () => {
+  const { setModalOpen, deleteEntry, entryId } =
+    useContext(ProductivityContext);
   return (
-    <div className="flex justify-center items-center overflow-x-hidden overflow-y-auto fixed inset-0 z-50 outline-none focus:outline-none backdrop-brightness-50">
+    <div className="flex justify-center items-center overflow-x-hidden overflow-y-auto fixed inset-0 z-50 outline-none focus:outline-none backdrop-brightness-75">
       <div className="relative w-auto my-6 mx-auto max-w-3xl ">
         <div className="border-0 rounded-lg shadow-lg relative flex flex-col w-full outline-none focus:outline-none">
           <div className="flex items-start justify-between p-5 border-b border-solid border-white rounded-t bg-black">
